fix: Update repository method signatures to use int32 for user IDs
- Changed HasFileAccess, HasFolderAccess userID parameter from uint to int32
- Changed GetFolderByID and related folder repository methods to use int32
- Added int32() type casts in folderUseCase.go where calling repository methods
- Updated all repository interfaces and implementations for consistency

+++ services/cloudRepositoryService/features/cloudRepository/repository/fileShareRepository.go
@@ -70,7 +70,7 @@ func (r *FileShareRepository) DeleteFileShare(ctx context.Context, fileID uint, 
 }
 
 // HasFileAccess checks if a user has access to a file (owner, shared, or folder shared)
-func (r *FileShareRepository) HasFileAccess(ctx context.Context, userID uint, fileID uint) (bool, error) {
+func (r *FileShareRepository) HasFileAccess(ctx context.Context, userID int32, fileID uint) (bool, error) {
 	// Get the file
 	var file entity.CloudFile
 	err := r.db.WithContext(ctx).
@@ -85,7 +85,7 @@ func (r *FileShareRepository) HasFileAccess(ctx context.Context, userID uint, fi
 	}
 
 	// Check if user is the owner
-	if file.UserID == userID {
+	if file.UserID == uint(userID) {
 		return true, nil
 	}
